requirements.txt
# transformers==4.45.2
# accelerate==0.33.0
# huggingface_hub==0.24.6
runpod==1.7.0
Pillow==10.4.0
transformers==4.46.3
accelerate==0.34.2
huggingface_hub==0.26.5